Hoang Anh them code

+++ src/main/java/com/example/webnghenhac/service/UserService.java
@@ -3,8 +3,12 @@ package com.example.webnghenhac.service;
 import com.example.webnghenhac.dto.request.UserCreationRequest;
 import com.example.webnghenhac.dto.request.UserUpdateRequest;
 import com.example.webnghenhac.entity.User;
+import com.example.webnghenhac.exeception.AppException;
+import com.example.webnghenhac.exeception.ErrorCode;
 import com.example.webnghenhac.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,6 +18,7 @@ import java.util.UUID;
 public class UserService {
     @Autowired
     private UserRepository  userRepository;
+
 
     public List<User> findAll() {
         return userRepository.findAll();
@@ -26,9 +31,14 @@ public class UserService {
 
     public User createUser(UserCreationRequest request) {
         User user = new User();
+        if(userRepository.existsByUsername(request.getUsername()))
+            throw new AppException(ErrorCode.USERNAME_EXISTED);
         user.setUsername(request.getUsername());
-        user.setPassword(request.getPassword());
+
         user.setEmail(request.getEmail());
+        user.setRole(request.getRole());
+        PasswordEncoder passwordEncoder = new BCryptPasswordEncoder(10);
+        user.setPassword(passwordEncoder.encode(request.getPassword()));
 
         return userRepository.save(user);
     }
@@ -42,6 +52,8 @@ public class UserService {
     }
 
     public void deleteUser(String id) {
+        if(!userRepository.existsById(id))
+            throw new RuntimeException("id does not exist");
         userRepository.deleteById(id);
     }
